Extract shared controller methods to controller concern
