+++ engines/order_management/app/controllers/order_management/reports_controller.rb
@@ -2,10 +2,12 @@
 
 module OrderManagement
   class ReportsController < Spree::Admin::BaseController
-    def show
-      render_report && return if params[:q].blank?
+    include OrderManagement::Concerns::ReportsActions
 
-      @report = report_class.new(spree_current_user, params[:q], params[:options])
+    def show
+      render_report && return if ransack_params.blank?
+
+      @report = report_class.new(spree_current_user, ransack_params, report_options)
 
       if export_spreadsheet?
         export_report
@@ -16,31 +18,9 @@ module OrderManagement
 
     private
 
-    def report_type
-      params[:report_type]
-    end
-
-    def report_subtype
-      params[:report_subtype]
-    end
-
-    def report_class
-      return if report_type.blank?
-
-      report_loader.report_class
-    end
-
-    def report_loader
-      @report_loader ||= Reports::ReportLoader.new(report_type, report_subtype)
-    end
-
-    def export_spreadsheet?
-      ['xlsx', 'ods', 'csv'].include?(report_format)
-    end
-
     def export_report
       render report_format.to_sym => @report.public_send("to_#{report_format}"),
-             :filename => filename
+             :filename => report_filename
     end
 
     def render_report
@@ -66,23 +46,6 @@ module OrderManagement
       @distributors = form_options.distributors
       @suppliers = form_options.suppliers
       @order_cycles = form_options.order_cycles
-    end
-
-    def form_options_required?
-      [:packing, :customers, :products_and_inventory, :order_cycle_management].
-        include? report_type.to_sym
-    end
-
-    def report_format
-      params[:report_format]
-    end
-
-    def filename
-      "#{params[:report_type] || action_name}_#{timestamp}.#{report_format}"
-    end
-
-    def timestamp
-      Time.zone.now.strftime("%Y%m%d")
     end
   end
 end
